Rename: gui/hardware_set/{device_connection.py => device.py}
