+++ finesse/gui/hardware_set/device.py
